requirements.txt
anyio==4.13.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
lxml==6.0.2
mercadopago==2.3.0
outcome==1.3.0.post0
pycountry==26.2.16
pycryptodome==3.23.0
PySocks==1.7.1
python-dotenv==1.2.2
python-telegram-bot==22.7
requests==2.32.5
sniffio==1.3.1
sortedcontainers==2.4.0
streamlink==8.2.1
trio==0.33.0
trio-websocket==0.12.2
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
